fix: Integer -> Long nos IDs e instanciao do center no sedding e relacionamento estabelecido com o packageProduct

+++ src/main/java/com/development/order/model/entities/Client.java
@@ -23,7 +23,7 @@ public class Client implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	private Integer cpf;
 	private Integer cep;
@@ -35,7 +35,7 @@ public class Client implements Serializable {
 
 	}
 
-	public Client(Integer id, String name, Integer cpf, Integer cep, Instant DataBuy) {
+	public Client(Long id, String name, Integer cpf, Integer cep, Instant DataBuy) {
 		this.id = id;
 		this.name = name;
 		this.cpf = cpf;
@@ -46,7 +46,7 @@ public class Client implements Serializable {
 	public List<PackageProduct> getPkgs(){
 		return pkgs;
 	}
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
@@ -62,7 +62,7 @@ public class Client implements Serializable {
 		return cep;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
