+++ src/main/java/com/development/order/repositories/PackageProductRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.development.order.model.entities.PackageProduct;
 
-public interface PackageProductRepository extends JpaRepository<PackageProduct, Integer>{
+public interface PackageProductRepository extends JpaRepository<PackageProduct, Long>{
 
 }
